day1: complete Java basic exercises

+++ src/Day1/DataType.java
@@ -4,7 +4,7 @@ public class DataType {
     public static void main(String[] args) {
         String name="张三";
         int age=20;
-        double height=1.75;
+        float height=1.75f;
         boolean bool=true;
         System.out.println("姓名："+name+"  "+"年龄："+age+"  "+"身高："+height+"  "+"是否学生："+bool);
         System.out.printf("姓名：%s  年龄：%d  身高:%.2f  是否是学生：%b",name,age,height,bool);
